api/requirements.txt
Flask
pymongo
markdown
pyjwt